tests/wpt/web-platform-tests/wai-aria/role/basic.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<div id='d' style='height: 100px; width: 100px' role="group" aria-label="test label"></div>
<h1 id="h">test heading</h1>
<script>

promise_test(async t => {
  const role = await test_driver.get_computed_role(document.getElementById('d'));
  assert_equals(role, "group");
}, "tests explicit role");


promise_test(async t => {
  const role = await test_driver.get_computed_role(document.getElementById('h'));
  assert_equals(role, "heading");
}, "tests implicit role");

</script>
